requirements.txt
google-api-python-client==2.123.0  # YouTube Data API
openai>=1.0.0
python-dotenv==1.0.1               # Loads .env secrets
requests>=2.31.0                   # Common helper (if you use utils.py)
